fix pg type

+++ src/api/credential.ts
@@ -18,7 +18,7 @@ export const makeCredentialPersistence = () => ({
     validateId,
     T.chain(encodeId),
     T.chain(({ id }) =>
-      query(`SELECT * FROM "public"."credentials" WHERE "id" = $1::bigint`, id)
+      query(`SELECT * FROM "public"."credentials" WHERE "id" = $1::integer`, id)
     ),
     T.chain((_) =>
       _.rows.length > 0 ? T.succeed(_.rows[0]) : T.fail(new CredentialNotFound())
